+++ CHANGELOG.md
@@ -9,10 +9,14 @@
 * feature
   * Add generator for configure csv2hash on Rails app
 
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/16)
+
 ### VERSION 0.6.4
 
 * feature
   * make configuration file can be writing with ERB interpolation.
+
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/15)
 
 ### VERSION 0.6.3
 
@@ -22,15 +26,21 @@
 * feature
   * add yml loader
 
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/14)
+
 ### VERSION 0.6.2
 
 * feature
   * Auto discover, you can add floating position on mapping rules
 
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/12)
+
 ### VERSION 0.6.1
 
 * deprecations
   * Structure validation rules, MinColumn, MaxColumn are replaced by :min_columns, :max_columns
+
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/11)
 
 ### VERSION 0.6.0
 
@@ -40,6 +50,8 @@
 * refactoring
   * replace arguments for rules by DSL of cells
 
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/10)
+
 ### VERSION 0.5.0
 
 * backwards incompatible changes
@@ -47,6 +59,8 @@
 
 * refactoring
   * remove params ``` ignore_blank_line ``` to benefit of options of Hash
+
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/9)
 
 ### VERSION 0.4.0
 
@@ -56,6 +70,8 @@
 * refactoring
   * remove params ``` break_on_failure ``` to benefit of ``` Csv2hash::Main#parse! ``` of ``` Csv2hash::Main#parse ```
 
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/8)
+
 ### VERSION 0.3.0
 
 * backwards incompatible changes
@@ -64,6 +80,8 @@
 
 * refactoring
   * Use adapter to select source of data, either file.csv or Array of data, is more transparently, not impact signature of Csv2Hash constructor
+
+* [fullchanges](https://github.com/FinalCAD/csv2hash/pull/7)
 
 ### VERSION 0.2.1